--- app/src/main/java/com/aconno/sensorics/AdvertisementFormatReader.kt
@@ -1,55 +1,0 @@
-package com.aconno.sensorics
-
-import android.content.Context
-import android.content.res.AssetManager
-import com.aconno.sensorics.device.format.AdvertisementFormatReader
-import com.aconno.sensorics.model.GenericFormatModel
-import com.google.gson.FieldNamingPolicy
-import com.google.gson.Gson
-import com.google.gson.GsonBuilder
-import java.io.IOException
-
-class AdvertisementFormatReaderImpl(
-    private val applicationContext: Context
-) : AdvertisementFormatReader {
-
-    companion object {
-        private const val PATH = "formats"
-    }
-
-    private val gson: Gson = GsonBuilder()
-        .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
-        .create()
-
-    override fun getPreloadedFormats(): List<String> {
-        val assets = applicationContext.assets
-
-        val results = mutableListOf<GenericFormatModel>()
-        val files = assets.list(PATH)
-
-        files?.forEach {
-            try {
-                val fileData = getFileData(assets, "$PATH/$it")
-                val advertisementFormat = toAdvertisementFormat(fileData)
-                results.add(advertisementFormat)
-            } catch (e: IOException) {
-                e.printStackTrace()
-            }
-        }
-
-        return results.map { gson.toJson(it) }
-    }
-
-    private fun toAdvertisementFormat(data: String): GenericFormatModel {
-        return gson.fromJson(data, GenericFormatModel::class.java)
-    }
-
-    private fun getFileData(assetManager: AssetManager, fileName: String): String {
-        val inputStream = assetManager.open(fileName)
-        val size = inputStream.available()
-        val buffer = ByteArray(size)
-        inputStream.read(buffer)
-        inputStream.close()
-        return String(buffer)
-    }
-}
