+++ app/src/main/java/com/aconno/sensorics/dagger/application/AppModule.kt
@@ -1,11 +1,9 @@
 package com.aconno.sensorics.dagger.application
 
 import android.arch.persistence.room.Room
-import android.bluetooth.BluetoothAdapter
 import android.content.SharedPreferences
 import android.preference.PreferenceManager
 import android.support.v4.content.LocalBroadcastManager
-import com.aconno.sensorics.BluetoothStateReceiver
 import com.aconno.sensorics.IntentProviderImpl
 import com.aconno.sensorics.SensoricsApplication
 import com.aconno.sensorics.data.mapper.*
@@ -20,14 +18,9 @@ import com.aconno.sensorics.data.repository.publishdevicejoin.PublishDeviceJoinR
 import com.aconno.sensorics.data.repository.restpublish.RestPublishRepositoryImpl
 import com.aconno.sensorics.data.repository.sync.SyncDao
 import com.aconno.sensorics.data.repository.sync.SyncRepositoryImpl
-import com.aconno.sensorics.device.BluetoothCharacteristicValueConverter
 import com.aconno.sensorics.device.SmsSenderImpl
 import com.aconno.sensorics.device.TextToSpeechPlayerImpl
 import com.aconno.sensorics.device.VibratorImpl
-import com.aconno.sensorics.device.bluetooth.BluetoothImpl
-import com.aconno.sensorics.device.bluetooth.BluetoothPermission
-import com.aconno.sensorics.device.bluetooth.BluetoothPermissionImpl
-import com.aconno.sensorics.device.bluetooth.BluetoothStateListener
 import com.aconno.sensorics.device.notification.IntentProvider
 import com.aconno.sensorics.device.notification.NotificationDisplayImpl
 import com.aconno.sensorics.device.notification.NotificationFactory
@@ -47,13 +40,9 @@ import com.aconno.sensorics.domain.interactor.ifttt.UpdatePublishUseCase
 import com.aconno.sensorics.domain.interactor.repository.*
 import com.aconno.sensorics.domain.interactor.resources.GetFormatsUseCase
 import com.aconno.sensorics.domain.model.Device
-import com.aconno.sensorics.domain.model.Reading
-import com.aconno.sensorics.domain.model.ScanDevice
-import com.aconno.sensorics.domain.model.ScanResult
 import com.aconno.sensorics.domain.repository.DeviceRepository
 import com.aconno.sensorics.domain.repository.InMemoryRepository
 import com.aconno.sensorics.domain.repository.SyncRepository
-import com.aconno.sensorics.domain.scanning.Bluetooth
 import com.aconno.sensorics.domain.serialization.Deserializer
 import com.aconno.sensorics.domain.serialization.DeserializerImpl
 import com.google.gson.FieldNamingPolicy
@@ -73,44 +62,7 @@ class AppModule {
 
     @Provides
     @Singleton
-    fun provideBluetoothStateReceiver(bluetoothStateListener: BluetoothStateListener) =
-        BluetoothStateReceiver(bluetoothStateListener)
-
-    @Provides
-    @Singleton
-    fun provideBluetoothStateListener() = BluetoothStateListener()
-
-    @Provides
-    @Singleton
-    fun provideBluetooth(
-        sensoricsApplication: SensoricsApplication,
-        sharedPreferences: SharedPreferences,
-        bluetoothAdapter: BluetoothAdapter,
-        bluetoothPermission: BluetoothPermission,
-        bluetoothStateListener: BluetoothStateListener,
-        bluetoothCharacteristicValueConverter: BluetoothCharacteristicValueConverter
-    ): Bluetooth =
-        BluetoothImpl(
-            sensoricsApplication,
-            sharedPreferences,
-            bluetoothAdapter,
-            bluetoothPermission,
-            bluetoothStateListener,
-            bluetoothCharacteristicValueConverter
-        )
-
-    @Provides
-    @Singleton
-    fun provideBluetoothAdapter(): BluetoothAdapter = BluetoothAdapter.getDefaultAdapter()
-
-    @Provides
-    @Singleton
-    fun provideBluetoothPermission(): BluetoothPermission = BluetoothPermissionImpl()
-
-    @Provides
-    @Singleton
     fun provideInMemoryRepository(): InMemoryRepository = InMemoryRepositoryImpl()
-
 
     @Provides
     @Singleton
@@ -294,33 +246,6 @@ class AppModule {
         formatMatcher: FormatMatcher,
         deserializer: Deserializer
     ) = GenerateReadingsUseCase(formatMatcher, deserializer)
-
-    @Provides
-    @Singleton
-    fun provideFilteredScanResult(
-        bluetooth: Bluetooth,
-        filterByFormatUseCase: FilterByFormatUseCase
-    ): Flowable<ScanResult> {
-        return bluetooth.getScanResults().filter { filterByFormatUseCase.execute(it) }
-    }
-
-    @Provides
-    @Singleton
-    fun provideDevice(
-        filteredScanResult: Flowable<ScanResult>,
-        generateScanDeviceUseCase: GenerateScanDeviceUseCase
-    ): Flowable<ScanDevice> {
-        return filteredScanResult.concatMap { generateScanDeviceUseCase.execute(it).toFlowable() }
-    }
-
-    @Provides
-    @Singleton
-    fun provideReadings(
-        filteredScanResult: Flowable<ScanResult>,
-        generateReadingsUseCase: GenerateReadingsUseCase
-    ): Flowable<List<Reading>> {
-        return filteredScanResult.concatMap { generateReadingsUseCase.execute(it).toFlowable() }
-    }
 
     @Provides
     @Singleton
