+++ app/src/main/java/com/aconno/sensorics/dagger/mainactivity/MainActivityFragmentsModule.kt
@@ -9,7 +9,6 @@ import com.aconno.sensorics.ui.devices.SavedDevicesFragment
 import com.aconno.sensorics.ui.dialogs.ScannedDevicesDialog
 import com.aconno.sensorics.ui.livegraph.LiveGraphFragment
 import com.aconno.sensorics.ui.readings.GenericReadingListFragment
-import com.aconno.sensorics.ui.sensors.SensorListFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -23,10 +22,6 @@ abstract class MainActivityFragmentsModule {
     @FragmentScope
     @ContributesAndroidInjector
     abstract fun bindSavedDevicesFragment(): SavedDevicesFragment
-
-    @FragmentScope
-    @ContributesAndroidInjector
-    abstract fun bindSensorListFragment(): SensorListFragment
 
     @FragmentScope
     @ContributesAndroidInjector
