+++ app/src/main/java/com/aconno/sensorics/dagger/mainactivity/MainActivityModule.kt
@@ -4,8 +4,6 @@ import android.arch.lifecycle.ViewModelProviders
 import com.aconno.sensorics.BluetoothStateReceiver
 import com.aconno.sensorics.SensoricsApplication
 import com.aconno.sensorics.device.permissons.PermissionActionFactory
-import com.aconno.sensorics.device.usecase.RemoteUseCaseRepositoryImpl
-import com.aconno.sensorics.device.usecase.RetrofitUseCaseApi
 import com.aconno.sensorics.domain.interactor.filter.FilterByMacUseCase
 import com.aconno.sensorics.domain.interactor.repository.DeleteDeviceUseCase
 import com.aconno.sensorics.domain.interactor.repository.GetReadingsUseCase
@@ -17,8 +15,6 @@ import com.aconno.sensorics.domain.model.Reading
 import com.aconno.sensorics.domain.model.ScanDevice
 import com.aconno.sensorics.domain.repository.DeviceRepository
 import com.aconno.sensorics.domain.repository.InMemoryRepository
-import com.aconno.sensorics.domain.repository.LocalUseCaseRepository
-import com.aconno.sensorics.domain.repository.RemoteUseCaseRepository
 import com.aconno.sensorics.domain.scanning.Bluetooth
 import com.aconno.sensorics.ui.MainActivity
 import com.aconno.sensorics.ui.readings.ReadingListViewModel
@@ -30,8 +26,6 @@ import com.aconno.sensorics.viewmodel.resources.MainResourceViewModelFactory
 import dagger.Module
 import dagger.Provides
 import io.reactivex.Flowable
-import retrofit2.Retrofit
-import retrofit2.converter.scalars.ScalarsConverterFactory
 
 
 @Module
@@ -185,25 +179,6 @@ class MainActivityModule {
         filterByMacUseCase,
         getUseCaseResourceUseCase
     )
-
-    @Provides
-    @MainActivityScope
-    fun provideRetrofitUseCaseApi(): RetrofitUseCaseApi {
-        val retrofit = Retrofit.Builder()
-            .baseUrl("http://playground.simvelop.de:8095")
-            .addConverterFactory(ScalarsConverterFactory.create())
-            .build()
-
-        return retrofit.create(RetrofitUseCaseApi::class.java)
-    }
-
-    @Provides
-    @MainActivityScope
-    fun provideRemoteUseCaseRepository(
-        retrofitUseCaseApi: RetrofitUseCaseApi,
-        localUseCaseRepository: LocalUseCaseRepository
-    ): RemoteUseCaseRepository =
-        RemoteUseCaseRepositoryImpl(retrofitUseCaseApi, localUseCaseRepository)
 
     @Provides
     @MainActivityScope
