--- device/src/main/java/com/aconno/sensorics/device/format/AdvertisementFormatReader.kt
@@ -1,7 +1,0 @@
-package com.aconno.sensorics.device.format
-
-
-interface AdvertisementFormatReader {
-
-    fun getPreloadedFormats(): List<String>
-}
