--- device/src/main/java/com/aconno/sensorics/device/usecase/model/RemoteUseCase.kt
@@ -1,7 +1,0 @@
-package com.aconno.sensorics.device.usecase.model
-
-data class RemoteUseCase(
-    val name: String,
-    val serverTimestamp: Long,
-    val html: String
-)
