--- domain/src/main/java/com/aconno/sensorics/domain/format/RemoteAdvertisementFormat.kt
@@ -1,3 +1,0 @@
-package com.aconno.sensorics.domain.format
-
-data class RemoteAdvertisementFormat(val id: String, val serverTimestamp: Long, val format: String)
