--- domain/src/main/java/com/aconno/sensorics/domain/repository/LocalFormatRepository.kt
@@ -1,16 +1,0 @@
-package com.aconno.sensorics.domain.repository
-
-import com.aconno.sensorics.domain.format.AdvertisementFormat
-
-interface LocalFormatRepository {
-
-    fun getAllFormats(): List<AdvertisementFormat>
-
-    fun getAllFormatIds(): List<String>
-
-    fun getLastUpdateTimestamp(formatId: String): Long
-
-    fun addOrReplaceFormat(formatId: String, timestamp: Long, formatJson: String)
-
-    fun deleteFormat(formatId: String)
-}
