--- domain/src/main/java/com/aconno/sensorics/domain/repository/LocalUseCaseRepository.kt
@@ -1,10 +1,0 @@
-package com.aconno.sensorics.domain.repository
-
-interface LocalUseCaseRepository {
-    fun saveOrReplaceUseCase(name: String, lastModifiedDate: Long, html: String)
-    fun getLastUpdateTimestamp(sensorName: String): Long?
-    fun getAllUseCaseNames(): List<String>
-    fun deleteUseCase(sensorName: String)
-    fun getFilePathFor(sensorName: String): String
-    fun moveUsecasesFromAssetsToCache()
-}
