--- domain/src/main/java/com/aconno/sensorics/domain/repository/RemoteFormatRepository.kt
@@ -1,9 +1,0 @@
-package com.aconno.sensorics.domain.repository
-
-import com.aconno.sensorics.domain.format.AdvertisementFormat
-import io.reactivex.Completable
-
-interface RemoteFormatRepository {
-    fun getSupportedAdvertisementFormats(): List<AdvertisementFormat>
-    fun updateAdvertisementFormats(): Completable
-}
