--- domain/src/main/java/com/aconno/sensorics/domain/repository/RemoteUseCaseRepository.kt
@@ -1,7 +1,0 @@
-package com.aconno.sensorics.domain.repository
-
-import io.reactivex.Maybe
-
-interface RemoteUseCaseRepository {
-    fun updateUseCases(sensorName: String): Maybe<String>
-}
