+++ domain/src/main/java/com/aconno/sensorics/domain/repository/ResourcesRepository.kt
@@ -1,5 +1,4 @@
 package com.aconno.sensorics.domain.repository
 
 interface ResourcesRepository : ConfigRepository, FormatRepository, UseCaseRepository,
-    MainScreenRepository {
-}
+    MainScreenRepository
